✨ feat: apply toast notifications
