+++ src/app/(with-header-sidebar)/mypage/_components/Modal.module.css
@@ -8,7 +8,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  z-index: 9999;
+  z-index: 1000;
 }
 
 .modal {
