+++ src/lib/boardService.ts
@@ -1,4 +1,5 @@
 import axiosInstance from '@/lib/axiosInstance';
+import useToastStore from '@/store/toastStore';
 import {
   UpdateDashboardRequestParams,
   CreateDashboardRequestBody,
@@ -18,14 +19,18 @@ export const updateBoard = async (
   id: string,
   { title, color }: UpdateDashboardRequestParams
 ) => {
+  const addToast = useToastStore.getState().addToast;
   try {
     const response = await axiosInstance.put(`/dashboards/${id}`, {
       title,
       color,
     });
+    addToast('변경되었습니다.', 'success');
     return response.data;
   } catch (error) {
-    throw error;
+    if (error instanceof Error) {
+      addToast(error.message, 'error');
+    }
   }
 };
 
@@ -45,9 +50,13 @@ export const createDashboard = async ({
 };
 
 export const deleteDashboard = async (dashboardId: string) => {
+  const addToast = useToastStore.getState().addToast;
   try {
     await axiosInstance.delete(`/dashboards/${dashboardId}`);
+    addToast('대시보드가 삭제되었습니다.', 'success');
   } catch (error) {
-    throw error;
+    if (error instanceof Error) {
+      addToast(error.message, 'error');
+    }
   }
 };
