+++ src/lib/cardService.ts
@@ -1,5 +1,6 @@
-import { TaskFormValues } from '@/app/(with-header-sidebar)/dashboard/[id]/components/CreateTaskModal';
+import { TaskFormValues } from '@/app/(with-header-sidebar)/dashboard/[id]/components/CreateCardModal';
 import axiosInstance from '@/lib/axiosInstance';
+import useToastStore from '@/store/toastStore';
 
 export const deleteCard = async (cardId: number) => {
   try {
@@ -15,8 +16,9 @@ export const createCard = async (
   dashboardId: number
 ) => {
   const { image, ...body } = data;
-
+  const addToast = useToastStore.getState().addToast;
   let url = null;
+
   try {
     if (image) {
       const formData = new FormData();
@@ -38,9 +40,12 @@ export const createCard = async (
       ...body,
       ...(url && { imageUrl: url }),
     });
+    addToast('카드가 생성되었습니다.', 'success');
     return response.data;
   } catch (error) {
-    throw error;
+    if (error instanceof Error) {
+      addToast(error.message, 'error');
+    }
   }
 };
 
@@ -50,8 +55,9 @@ export const updateCard = async (
   cardId: number
 ) => {
   const { image, ...body } = data;
-
+  const addToast = useToastStore.getState().addToast;
   let url = null;
+
   try {
     if (image) {
       const formData = new FormData();
@@ -72,8 +78,11 @@ export const updateCard = async (
       ...body,
       ...(url && { imageUrl: url }),
     });
+    addToast('카드가 수정되었습니다.', 'success');
     return response.data;
   } catch (error) {
-    throw error;
+    if (error instanceof Error) {
+      addToast(error.message, 'error');
+    }
   }
 };
